requirements.txt
Flask
requests
pynput
pyperclip
Pillow
pytesseract  # optional for OCR; requires Tesseract installed on system
openai       # optional, if using OpenAI
